test: extract mock helpers for property-based testing
- Create mock-text-document.ts with MockTextDocument interface and create_mock_document factory function
- Create mock-handler-deps.ts with create_mock_handler_deps factory for HandlerDependencies
- Export new mock utilities from helpers/index.ts for reuse across property tests
- Replace inline mock HandlerDependencies creation in lsp-working-directory-option-server-response.prop.test.ts with create_mock_handler_deps call
- Replace inline mock HandlerDependencies creation in send-to-stata-statement-detection.prop.test.ts with create_mock_handler_deps call
- Normalize line endings in helpers/index.ts (CRLF to LF)
- Reduces code duplication and improves maintainability of mock objects across property-based tests
